backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opensearch-py==2.4.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1
websockets>=12.0
